Changes to secondary index creation
Minor adjustment to creation of partial indexes. Changed merging of partial indexes to require O(n) memory space where n is the number of files needed to merge, so the full index is never loaded into main memory. Adjusted the merging algorithm to be O(n) with respect to the number of characters.